Fixed: On opening the file, set tree horizontal scroll bar to the left border
When the name of any of the nodes is longer than the width of the tree panel, the horizontal scroll bar is displayed.
By default, if the selected note in the tree also has a longer node name, the horizontal scroll bar was set to the right edge,
causing the node hierarchy structure to not be clearly displayed.