Fixed: (IMPORTANT). "Show word count in status bar" in notes with 30000+ chars can corrupt that note
Typing in a note with more than 30000 characters, instead of showing (...) W:...... in the status bar,
would corrupt the content of the note, as indicated in #695.

It was caused by a seemingly innocent error (an oversight) in the first commit of the refactoring, in version 1.9.3,
but with important consequences.
In the affected code, "Text" was a property of a control of status bar. After refactoring it was the content
of the editor... It should have been written as "txt", not Text.