* Added a new INI option: URLWebDecode to adapt URLs with %XX characters
If True (1) it will manage %XX in URL as UTF8, finally converting the whole URL to ANSI or UTF8
depending on current codepage

Example:
https://ru.wikipedia.org/wiki/%D0%92%D0%B5%D0%B3%D0%B0,_%D0%9B%D0%BE%D0%BF%D0%B5_%D0%B4%D0%B5
->
https://ru.wikipedia.org/wiki/Вега,_Лопе_де

Certain characters that could be encoded with %XX won't be converted:

  ' ', '/', '?','!','''', '&', '%', '#', '$', '[', ']', '(',')', ',', ';', '*', ':', '@', '=', '+'

Ex: http://www.example.com/space%20here.html won't be modified

On "Insert URL" and "Choose Action for URL" dialogs, although URLWebDecode=1, you can force the URL not to be
modified pressing Shift when exiting URL field, or when clicking on OK (Insert URL) or Modify (Choose Action for URL).
Even if URL is not modified (because of Shift or URLWebDecode=0), when text URL is empty or equal to URL it will be set
with the decoded version of the URL field.
So, for example, if you pressed Shift while changing focus from "URL" to "Text", the URL field will maintain the %XX
characters, but the "Text" field will be set decoded. In the example, this way you can automatically have:
   URL: https://ru.wikipedia.org/wiki/%D0%92%D0%B5%D0%B3%D0%B0,_%D0%9B%D0%BE%D0%BF%D0%B5_%D0%B4%D0%B5
  Text: https://ru.wikipedia.org/wiki/Вега,_Лопе_де

When using ClipCap or Web Copy (Ctrl+W or Ctrl+Shift+W), the clip URL will be automatically adapted if URLWebDecode = 1.
Any other pasted hyperlinks, interspersed in the text, will be adapted only by opening and modifying via the
"Choose Action for URL" dialog box.

Ref: #618