Added new tree filtering
Buttons of "Tree toolbar" are now part of the tree panel, where a new textbox have been added.
This textbox allows to filter the nodes based (for now) on its name, with the configuration:
  Match case: False, Whole words only: False, Search hidden nodes: False, All the words

This tree filter can be combined with filtering based on the search panel (Find All)
Tree panel filtering (based on the search panel and/or tree textbox) can be temporarily disabled
(while highlighted the results) with "Filter nodes" button.