Fixed: Scrollbar position could be forgotten when returning to app with Hotkey (Ctrl+Shift+...) Ref: #720
Fixed a strange behavior of the RichEdit control when the application is restored through a Hotkey
of the form CTRL+SHIFT+...
The editor ended up receiving an EN_CHANGE message and the Change method of the control was called.
"Modified" property of the editor was changed to True (at least in W10, not in W11), and the position
of the scrollbar was lost, thus displaying the cursor position.

With another type of Hotkey combination it didn't happen (e.g. ALT+F7)