Note Editor refactoring -> TKntNoteUI
Created a new class (TKntNoteUI) to tend to centralize the UI management of a note.

For the moment it will still assume that there is only an entry in the note,
but later this class will include new funcionality related to the existence
of several entries in a note, tags, related resources, etc.

What has been already included is a top panel showing the name of the note and the creation date
(of the entry --for now equal to the creation of the note).
This panel is similar to the one requested in issue #704 (Node Information Panel)

The name of the note can be modified directly from that top panel.
The hints on the name an creation date display more information: path of the node and instant
of last modification in the note.

- Fixes in serialization of several note and note entry metadata.