New: It is allowed to toggle between RTF and Plain text only in active note
A new option has been added to the editor's context menu
It should be noted that the "Plain text only" option defined at Folder is now "*Default* plain text only". New notes in the folder defaults to that value. It is possible to combine RTF and Plain text notes in the same folder.