New: Two optional columns added to the tree panel: 'Flagged' and 'Date' (Note creation date)
A new option has been added to the tree context menu, to be able to mark the selected node(s)
as Flagged. It is also possible to toggle the flagged mark by clicking on the column corresponding
to the node.
In addition to highlighting in the context menu if a node is Flagged, an icon is also displayed
in the corresponding column (if visible).
The new columns can be made visible from F4 or F6 (Folder properties..), one or both.
Once visible, it is possible to reorder them. The resulting position will also be saved in the file.

Additional adjustments related to Hot Color

The order of the tabs in Folder properties has been swapped, to match the layout in the app. Now:
 Tree settings, Editor settings