Added option to tree panel context menu to filter out unflagged nodes
  View / Filter -> Filter out unflagged

This option is considered in conjunction with the existing tree filter and the search filter (if any).

Also: if you change the Flag on the selected node with CTRL+click (in the tree column or in the menu entry, 'Flagged'),
the 'Filter unflagged' will also be toggled on or off.
This way, for example, if there are no flagged nodes, CTRL+clicking on the marked column (if visible) of a node will
mark the selected node as marked and hide all other nodes.