'Tree filtering': It is possible to use the last modification date, with a simple syntax
It can be useful to be able to filter the tree by showing nodes that were modified (or created) during the current day
or since the last X days. This can be done very easily by specifying in the first characters of the search string
an expression of the form "-<number days back>"

Example. If current day is '15/10/2024' then:
  '-2' is equivalent to [Last Modified] >= '13/10/2024'
  '-0' is equivalent to [Last Modified] >= '15/10/2024'  (modified on current day)
  '-0 something'  -> nodes modified on current day that includes 'something' in their names (ignoring case)

   The filter text box will remind us of the date considered, showing a hint of the format:
     Last modified >= "15/10/2024"

If enclosed in "" it will managed as a normal text pattern:
 '"-2"' will look for '-2'

This will also work with Find All.

* Also, running Find All without adding any text is allowed, as long as some date criteria is set