Added new search option, saved in .INI: SearchPathInNodeNames (default False)
Allows to indicate whether when searching in the name of the nodes only the node itself should be considered
or the complete path should be used, that is, the names of the ancestors.
Until now, the search has always been done exclusively in the name of the node itself.

This new option can be set from the Find All panel, and will be used both in these searches and in the
filters made from the tree itself (Tree filtering)

I have included this option because, especially now that it is possible to search or filter based on the dates
of creation or last modification of the nodes, it is interesting to be able to combine the date criterion
with a filter based on the name of the nodes that allows selecting nodes based on its own dates (creation and
or last modification) and the names of their parents.
For example, let's suppose the following hierarchy of nodes, where we want to filter using Tree filtering (*)

Project_A
  Subject1
    ToDO
    Child A
	..
  Subject2

Project_B
    ToDO
    Issue New
	..

where nodes 'ToDO' have been modified today, and 'Child A' has been created today.

Searching for '-0' would select both 'ToDO' nodes and the 'Child A' node. If we wanted to focus on everything
related to "Project_A" we could search for '-0 Project_A', which would select the first ToDO node and the 'Child A' node.
If we searched for '-0 Project_A TODO' (or '-0 ToDO Project_A' for example), only the first 'ToDO' node would be selected.
This is because this way, considering the path (with the names of the parents), the selected nodes
would meet both criteria, date and text.

(*) From Find All it would be similar, taking into account that this new option will only be considered if
the selected search scope is ‘Only node name’ or ‘All’. From Find All it is also possible to configure other search criteria:
match case, whole word only, search type (exact phrase, all the words, any of the words). In tree filtering these criteria
are set to: match case = false, whole word only = False, search type = all the words