* Added new filtering option, saved in .INI, ShowChildren (show children of matching nodes) (default False)
When filtering, nodes that are matching are highlighted in light yellow (+). A parent node may be visible
because one of its children is a match, but it may also be a match itself, and be highlighted
in yellow.
By default, with the ShowChildren property set to False, when filtering the tree only the nodes that
satisfy the conditions (highlighted in yellow) are shown, as well as their parents.

If the ShowChildren property = True, the children of all matching nodes will also be made visible
(highlighted in yellow).
This allows, for example, to select one or more nodes using Filter Out Flagged, while at the same
time leaving their child nodes visible as well.

This makes it possible to obtain a functionality similar to what in some applications is called 'Hoist',
allowing to focus on a specific part of the note list tree.

  (+) The description corresponds to Tree Filtering. If the filtering is done from Find All, the matching
      nodes are highlighted with a blue font. But the above applies equally: depending on that option, the
	  children of these matching nodes will be shown or not.

* Added options to tree panel context menu:
    View / Filter -> Use path of nodes
                  -> Show children

  If either of these two options is changed and the current folder tree is filtered, it will be reapplied
  with the new options set.