Added accessibility to the Virtual TreeView control (used in tree panel, file manager and config.options)
Used a unit called VirtualTrees.Accessibility that implements IAccessible interfaces for the VirtualTree visual
components and the currently focused node.
To be able to read and recognize not only focused node but also any of the nodes when we place the cursor
over them, I have modified that unit: ui\VirtualTrees.Accessibility_MOD.pas

This allows the NVDA screen reader to read and recognize the tree nodes when you place the cursor over them,
and also when you go down or up the node tree with the arrows.

Ref: https://github.com/dpradov/keynote-nf/discussions/739#discussioncomment-11063013