(Accessibility) Added new .INI option, to facilitate reading Hints through the status bar
It is not the best solution but for now I have not managed to find a reasonable way to offer through
accessibility (IAccessibity) the hints that the application shows through the multiple controls that
it has.
For now and given that when the status bar is visible the hints are shown in one of its panels,
and having verified that readers such as NVDA offer key combinations to read the content of the status
bar, I have seen it convenient (using the new option) to make all the text panels of the status bar,
except for the one corresponding to the Hint, be shown empty, so that the screen reader can directly
read the text (the hint) without having to previously read all the other panels, which include the file
name, cursor position, file modification status, etc.
For example, the NVDA reader allows reading the status bar with the combination [NVDA]+End.

The option is "HintsAccesible". Just add it to the .INI file in the [KeyOptions] section.
If it is not added (default situation), or is set to zero, the behavior will be the same as always.
To obtain the new behavior, it must be added and set to 1. Ex:

[KeyOptions]
HintsAccesible=1

NVDA (Non-Visual Desktop Access) allows blind and vision impaired people to access and interact with
the Windows operating system and many third party applications.