* Find All: Revised the way 'All the words' search type works  (and some optimizations)
Within the search in a note (Node) KNT will change internally to 'Any of the words' mode from
'All the words' after the first match to allow locating and offering more useful results.
Let's suppose that we search for something like: [w1..w2 w3] (or simply [w1 w2 w3]
and we do it in a note with a text of the form:
                               ... w2 ... w1
                               ....... w3 ...
                               ...
                               ... w1 ... w2 ...

So far we have only been returning as a match the one that includes from the first w2
to w3. Since there are no more w3, the final terms, w1 and w2 are ignored because from the
final position of the last match (w3) we cannot find all the terms again.
But why are the first ones (w2 .. w1) going to be more significant than the second ones (w1 .. w2)?
We will surely be interested in knowing about the latter as well.

The nature of the 'All the words' method is maintained in the sense that we are offering all
these matches within this note because all the searched terms are found in it.
From that moment on, and once verified, we are interested in showing all the terms in the same
way as we do with 'Any of the words', including several terms together in the same match if
they are close and can be offered in the same extract.

If the text were as follows, we will still return the two pairs w1, w2.
The number of results may vary slightly depending on whether or not the word w3
has been displayed in any previous result.

                               ... w2 ... w1
                               ...
                               ... w1 ... w2 ..
                               .... w3 ...

- Also: Incorporated some improvement or optimization in the search process as well as in the identification of the extract to be offered