Change in behavior of ESC in the Scratch window
Pressing ESC in Scratchpad editor will minimize the application and not return focus to note editor, *if* [Configuration Options|General settings|Program hotkey | On Escape key = Minimize Main Window]. If 'On Escape' is not configured in that way, pressing ESC will focus the note editor, as before.

See the reasons in discussion "Pressing ESC in Scratchpad editor" (#744)
You can indicate whether you think this change is correct or not in the poll provided.