Fixed: Internally, the ID associated with each note entry is not being 0, as it should be.
This has no impact at the moment, other than a few more bytes being written to the .knt file for each note.
It will be fixed with the next version and the .knt file will no longer write those unnecessary values.

Since version 2.0 the model supports multiple "entries" for each note. Each entry has an ID, which starts
at zero. However, both files created with previous versions are converted to notes with a single entry,
and a single entry is still possible to create since version 2.0.0 Beta 8.
If yes, by design, all entries created should have 0 as their ID and therefore that ID would not
be recorded in the .knt file, which is the default value. And it would also not be necessary to
indicate at note level, and together with the cursor position, the ID of the selected entry.
But due to an error that did not appear in the executable generated in debug mode (but did in release mode),
the adaptation to the new format of files created with older versions has been assigning an undefined value
(which has been limited to one byte). Thus, something like the following can be seen in the created files:

%*
GI=2
ND=Plain text node
LM=2511242205
NS=02
SE=96
SS=33
%.
id=96
DC=2511242204
NS=0002
%>
;A line in the Plain text note...

Where it really should have been:

%*
GI=2
ND=Plain text node
LM=2511242205
NS=02
SS=33
%.
DC=2511242204
NS=0002
%>
;A line in the Plain text note...

because the entry id should have been 0 (id=0)

And although after the correction the entries are now numbered correctly starting from 0, the values
recovered when reading are respected. And this would imply that the following entries would continue
with an ID consecutive to the one read.

As in practice only one entry is still used, the solution will be that in the loading process
it will ignore those "id" and assume 0 when created with the a model version <= 3.0

When it is possible to manage multiple entries per note, the version will be increased.