Fixed: Changed 'NoteState' token ("NS" -> "Ns")
This will force any "NS" tokens loaded as 'NoteState' to be discarded and not saved for now.
It does not affect files saved up to now with versions 2.0.0 Beta because in these versions
the NoteState values are not yet being used.
On the contrary, some of those defined in NoteEntryState, which was also configured as "NS",
are being used.

Rearranging tokens (constants) in kn_Const.pas