The functionality now available from kntutils.dll is included in keynote.exe
This will increase the .exe file by just under 120kb, but at the same time it will no longer be necessary
to use kntutils.dll (2491Kb).
The included code is used to be able to configure the keyboard shortcuts as well as perform conversions between HTML and RTF
(in both directions), depending on the selected import/export method.

This also makes it easier to perform the translation of the keyboard shortcut configuration window using the
existing mechanism.

(The inclusion is done through a compilation constant. It may be changed in the future)