Fixed: Some words marked as target for bookmarks could be [partially] ignored when removing hidden marks (ex: exporting to RTF)
Besides when exporting to RTF or HTML, it also could be detected: copying a text including that bookmark
and pasting it into another program. In the process of removing hidden bookmarks, a word (or fraction)
could be ignored.
For example, in note "Image management in KeyNote NF" in Help, if we copied the text that
includes "Compression and encryption" (selecting also text before), and paste outside of it
the text pasted was "ssion and encryption" instead of "Compression and encryption"