Export.. KNT internal links without marker are also converted to standard RTF format (works with MS Word and LibreOffice)
These are links created with versions of KeyNote prior to 1.8.0, where markers were not used
to identify the destination.

Once converted, the target they point to will remain stable, even if the RTF document is modified
(from Word, LibreOffice or similar, never from WordPad). However, the position will be the same as the one recorded
in KeyNote (and to which the link currently jumps), which might not point to exactly the same word as it did when
it was initially inserted, precisely because it was based exclusively on the position and not on the use of markers,
as current links do.