* Improvements in Export...: Sections, page header, optional saving of each folders to a new file, ...
Optional use of sections is included:

By using sections it is possible to create groupings of notes with a defined maximum depth level. Seen another way, each time a note with a depth equal
to or less than the set value is processed a new section is started. If you enable the use of sections you will be able to:
- Decide whether the section will start on a new page or be saved in a separate file (one of two alternatives)
  (When printing, the only possible option, automatically selected, will be to start each section on a new page)

- Set a section-dependent page header
  By default, the name of the highest level node on which the node depends (and which is included in the export) will be used as the page header.
  It is possible to define, through the option "Top N levels as Page Header (in sections)", the level of detail in the header, adding the name
  of other lower level parent nodes.
  The value set for the number of levels in the page header may be less than or equal to the depth indicated for the section

The use of page headers is available both when exporting to RTF and when printing and previewing.

In current version, when exporting keynote folders, you could choose to put all nodes to a single target file, or to export each tree node
to a separate target file. Now it is also possible with the use of sections; you only have to define a dS value equal or higher than the level of
the deepest node you want to consider.

Other options included:
- Start each Note o a new Page
  Allows all notes to start on a new page. This option is only available for RTF export and printing, and can be activated independently of using
  sections or the "Save each Folder to a new File" option.

- Save each Folder to a new File
  When you include notes from more than one folder as the export source, if you activate the "Save each Folder to a new File" option, a file will
  be generated for each folder to be exported. Otherwise, all of them will be exported to a single file (unless sections are being used and these
  are saved to disk)..
  This option can only be selected if sections are not used or are not configured by setting "Save each Section to a new File".

* Also: Some fixes and optimizations in the conversion of KNT internal links
* Other improvements to the export process, such as the way file names are generated.