IMPORTANT. Fixed: Saving to disk is ignoring changes not yet committed, in other Folders editors
When saving, changes made on an editor of a non active folder, that had not yet been transferred to the model were being ignored. Changes in the editor are automatically transferred to the model when another node is selected. The problem could be more problematic for those of you who may be working on several folders with a single node.
Ref: #791, #792