Adaptations to allow 64 bit version of KeyNote NF
Some issues to remember:
* The HKEY_CLASSES_ROOT constant and other registry keys are defined in a way that they are not directly interpreted as simple integer constants on 64-bit architecture.
On Windows, constants like HKEY_CLASSES_ROOT are defined with a value that uses the upper part of the address space, which causes the compiler to treat these constants as NativeInt instead of Integer when compiling for 64-bit, resulting in error E2026: Constant expression expected.
The expression HKEY(Integer($80000000)):
On Delphi on 64-bit systems, these types of constants can cause problems because NativeInt on 64-bit is a 64-bit integer, while the expected values in the case are 32-bit.
The TRootKey type is an enumerated type:
The case expects constants of type HKEY_* to be compared directly with constant values that can be mapped to TRootKey values, but due to the size of constants in 64-bit, the error occurs.
=> Force explicit conversion to Integer on comparison: In the 64-bit context, you can force the value of FRegistryKey to be treated as an Integer to match the 32-bit constants expected in the case.

* Error en P := SmallPointToPoint(TSmallPoint(lParam));  (P: TPoint)
TSmallPoint is a structure that represents two SmallInt values (16 bits each) in a 32-bit register. It is commonly used to pack point coordinates (x, y) into a single LongInt or LParam value in Windows messages.
lParam is a 32-bit parameter in Windows messages, which on a 32-bit architecture is the same size as a LongInt (4 bytes). On a 64-bit architecture, lParam expands to 64 bits (NativeInt), leading to potential inconsistencies when trying to cast it directly to TSmallPoint.
In a 64-bit environment:
lParam is a 64-bit value, while TSmallPoint is still 32-bit.
Directly casting TSmallPoint(lParam) may result in an error because you are trying to cast a 64-bit value to a 32-bit value, resulting in data loss.
=> The solution is to ensure that only the first 32 bits of lParam are used to construct the TSmallPoint. This can be achieved by using an explicit and safe cast to Cardinal or LongInt before conversion to TSmallPoint. This ensures that only the low 32 bits of lParam are considered.
P := SmallPointToPoint(TSmallPoint(Cardinal(lParam)));

* To enable debugging:
    32 bit: compiler debug information enabled.
    64 bit: compiler **and linker** debug information enabled.

  Ref: https://stackoverflow.com/questions/30171572/unable-to-debug-64-bit-application-in-release-configuration-with-debug-informat

* c00000005 ACCESS_VIOLATION exceptions produced in calls to SendMessage, related to the handling of pointers and data types in the transition from 32-bit to 64-bit:
In 32-bit, Integer and Longint are 4 bytes long, while in 64-bit the WPARAM and LPARAM parameters used in SendMessage are 8 bytes long (type NativeInt or NativeUInt). If Integer or Longint is used to pass a pointer, this can cause an access violation because the pointer size is not being handled correctly.