Optimization of the caching process for data used to perform searches
When loading the file, all notes are scanned and a special plain text version is obtained that corresponds to the positions in the editor,
which makes it possible to perform searches on text strings and not through calls to the RichText control (for which it is first necessary
to load a RichText control with the content of the note).
In order not to affect the use of the file, this process is only performed while there is no user activity. Until now, this process
was performed on a small number of notes at a time, stopped if that number was reached or user activity was detected, and resumed
only in response to the configured timer (triggered every 2 seconds). This meant that in files with an extremely high number of notes
the process could take forever.
The process has been revised so that all notes can be scanned at once, while there is no user activity.

In addition, the way in which this data is obtained has been significantly optimized in the case of flat notes (not RTF). In these cases,
instead of using the RichText control, the necessary text is obtained directly from the content of the note, with string processing.
In this way, the time required to obtain this data (in text-type notes) can be divided by 3000.
As an example:

     // The first two strings used for comparison contained ANSI text, small in size. The first in UTF8
     // was small, and the second occupied 15481 bytes, with characters in multiple languages, along with Non-Emoji Symbol,
     // with symbols from:
     //   https://www.lennyfacecopypaste.com/text-symbols/square-rectangle.html
     //   https://www.prepressure.com/fonts/basics/emoji-list
     //    ...

     // N: Number of iterations. Time in seconds

     //3º:   N:300.000 ->   0,125   | 0,42199  | utf8:   0,15600 |      6,85900
     //2º:   N:300.000 -> 374,0697  | -        | utf8: 428,90    | 20.568,9

     // The most optimal option is the third one, using MemoryStreamToString, TryUTF8ToUnicodeString and StringReplace
     // Compared to it, getting it using RichText, with calls to LoadStreamInRTFAux and RTFAux.TextPlain
     // is many orders of magnitude slower. Almost 3000 times slower:

     //3º:   N:300.000 ->      1   | -        | utf8:      1    |        1
     //2º:   N:300.000 -> x 2992   | -        | utf8:  x 826    |   x 2998

     With a very small ANSI string, processed 300,000 times, the new method takes 0.125 seconds, while the old method
     took over 6 minutes.
     If the string was UTF8 of a medium size, 300,000 iterations could take almost 7 seconds versus 5.7 hours...

    (The time shown for using RichText is estimated from that obtained with N=10.000 on a small ANSI string
     and N=1000 on the indicated UTF8 string)