Fixed: Access violation when deleting or certain other actions on new/completely empty nodes (Ref #800)
The exception is caused by certain actions with the presence of completely empty nodes, which never had any data.
These empty nodes are, for example, new nodes created, before you write something to them. After writing something,
deleting the node content does not cause this problem.

The identified actions that could raise the exception are:

- Deleting node
- Removing a folder containing that node
- Doing a Merge with a file containing that empty node
- Applying Toggle Text / RTF on the node
- Changing the image storage mode of the file (containing that empty node)

This was a bug present in both 32-bit and 64-bit versions, but it does not normally appear in the 32-bit version.
Memory management in the 32-bit build seems to be much more permissive.