Working on Fold and Unfolding, with the identification of blocks, and its possible use with tags
Some fixes and tests of the use of opening and closing word/token pairs, now defined manually but later configurable in the application.
Any explicitly defined or inferred block can be collapsed and unfolded (compacted and expanded).

A block can be folded by double-clicking on a word, which will act as an opening, compacting it to the position identified as closing for it. It is also possible to fold a block from a manual selection (which can include other folded blocks, but always completely, not partially) and using the "Fold" option in the editor's context menu.

When a block is folded, a hyperlink with the text "+" is displayed, and it is closed with "..." and a special character at the end.
The folded block can be copied, cut or even deleted while it is fully selected, but it cannot be edited or deleted while folded; it is protected text.
A folded block can be expanded by clicking on the "+", or by selecting "Unfold" from the editor's context menu, with the cursor positioned inside it. If it included nested blocks, they will become visible, but will remain folded.

* It is possible to fold blocks defined from pairs of words or tokens, one opening and one closing. Examples:
   IF / END IF
   ** / **
   <> / </>

  If the two words are different, this allows nested blocks to be defined, and therefore kept folded.

* It is also possible to define tags (eg. "#ToDO")
If it is written immediately before a line break or ends in ":" (e.g. "#ToDO:") it is understood as an "open" tag, whose scope (the end of the block it affects) extends to a general closing chain of tags (in principle "##") or, if there is none, to the end of the note.
If it is not an open tag, its scope extends to the end of the paragraph in which it is included.

At the moment, these opening and closing tokens are searched for in a case-sensitive manner.

* When you double-click on a word that is not defined as a block opening word nor as a tag, the end of the block to be folded is searched for as follows:
If the string "[.]" is found after that word, it will be folded to it.
If it is not found, the end of the paragraph in which it is located will be considered as the end of the block.

Relatedly, when a block is collapsed by manually selecting text, a "[.]" will automatically be added to the end of the created block. That string will not be visible when the block is collapsed, but it will be visible when it is expanded (by running Unfold or clicking on '+'). Clicking on that same initial word (or another word before [.]) again will collapse it back to the previous position.
At any time, if that is not of interest, you can remove [.], so that the generic blocks (without defined tokens) are folded up to the end of the paragraph.

ToDO: Consider other possible generic block endings, for bullet points, or highlighted texts, as titles or sections?