Automatic renaming of newly created tags and automatic deletion of tags created by mistake
While entering new tags in the text editor, it is possible to add incorrect ones by mistake.
For example, it is possible to enter "#Eror" for the first time instead of "#Error". It is very likely that the tag "Eror" was added
before we realized it. By removing the text and typing ok the name, we will add the correct tag (Error), but the first one (Eror) will be kept.

To avoid having to delete tags created by mistake, all new tags created through the editor, by typing the "#" character and then the name of a tag,
will initially do so temporarily (and will be highlighted in green in the selector and in the list of tags in the Tags tab). When the focus moves
from the editor to another control, the added tags that are still in the text will be automatically confirmed, and the rest will be deleted.
In the same way, if while creating new tags in this way we write a tag in a different way, it will be understood that it is being corrected,
and it will be automatically renamed. Once the creation of the tag has been confirmed, its name can only be modified from the list of tags,
in the Tags tab.
Eg. If you are adding it temporarily and you change its spelling, we assume that you want to correct it. Eg: "TOdo" and then "ToDO"

Exactly the same thing happens with adding tags added from the text control available in the note identification panel.