When deleting a tag, optionally allow deleting in-text references from all existing notes
When deleting a tag from the tag list, this tag is now removed from the notes' metadata.
Depending on a new option (to be implemented), AutoDiscoverTags, you may be asked whether to also remove,
from the text of existing notes, all references to deleted tags.

If AutoDiscoverTags=True, it makes no sense to ask whether or not to remove references within the text of notes,
as they would be automatically added back when saving the notes. They will therefore be removed, and this will
be indicated along with the initial confirmation on whether to delete the tag.
In the case that AutoDiscoverTags=False KeyNote will ask for confirmation. It could be that we do not want to
remove references (uses) to tags already applied, which could still be searched for.
In fact, it could be a way to make it easier to find certain 'sensitive' information, perhaps hidden in folded
text, tagged with a tag not included in the visible list of tags.