* Improved the tag selector window + Refactorization
The floating window no longer steals focus from the main window. Among other things, clicking on the editor now
when the selector is visible closes the selector and places the focus at the clicked position, and does not
simply return focus to the editor. The behavior is thus equivalent to that of other programs.
Extracted the selector window functionality into a class from which TTagSelector is now derived: TSelector.