New control in tree panel to allow the selection of multiple tags to filter nodes with.
It is allowed the introduction of both individual tags and parts of them, which allow the selection of several at a time.
In the latter case, they are treated as OR conditions.
Example:
Suppose that there are defined tags such as #PryA.Module1 #PryA.Module2 ... #Bug #Bug.High #Bugh.Low
If you enter: "PryA Bug.Low" then (PryA.Module1 OR PryA.Module2 .. ) AND (Bug.Low) will be considered.

If a term ends in '*' and it is not identified as a registered tag, it will be assumed that we want to
indicate the search for tags that contain the entered substring (not just the one that ends with it).
In this way, if we have for example the tags: "Error", "ErrorHigh", "ErrorLow" and we want to select only
the first one, we will write "Error" (or simply select it from the selector, in the many possible ways).
If we want to consider the three tags (an error, whatever the severity or if it has not been defined)
we will write "Error*".
When there is no match with any individual tag it is not necessary to add * to select several.
E.g.: If we have "PryA.ModuleA" and "PryA.ModuleB" we could write "PryA" to select the two
(in an OR condition)