Added two options to the context menu of Tags tab: "Filter Notes with selected Tags (Ctrl: OR)", "Clear Tag filter on notes"
The first entry will filter the notes tagged with ALL the selected items (AND). If Ctrl is pressed while clicked,
then it will filter all the notes tagged with ANY of the selected items (OR)

Note:
The tag filter text control of the tree panel allows a mix of 'AND' and 'OR' operands but only
if any of the words doesn't identify an unique tag, like "Err" to select Error OR Error.High
(for example). But it doesn't allow something like 'Bug' OR 'Error'. Currently even "OR" could be a tag (#OR).
Something like "Err ToDO" will be interpreted as "(Error OR Error.High) AND ToDO"

When you click "Filter Notes with selected Tags (Ctrl: OR)", the tag filter text control is populated with
the list of selected tag names and, if Ctrl was pressed, preceded by "OR:", like: "OR: Bug Error"
The correct filter to be executed is registered by this menu operation (Bug OR Error), but it not corresponds
to the filter that the text control will deduce if that names were typed together. So, currently, if after
applying this action with Ctrl you enter in the text control, when you exit it, it will be probably changed
to "Bug Error" (because OR won't be identified like a registered tag) and interpreted as an AND operation
("Bug & Error")