Improvements to the Tags tab and tag filtering
Removed the context menu options "Filter Notes with selected Tags (Ctrl: OR)", "Clear Tag filter on notes".
These have been replaced by a checkbox "Show use of tags in tree" and a combo "Filter notes with ALL/ANY tags"
If "Show use of tags in tree" is checked, then only tags applied as metadata to any note will be displayed,
with an indication of the number of nodes containing them.
Then, as we select one or more tags, and depending on the filtering mode (All or Any), the nodes of the
tree will be filtered and the available tags and their associated number of nodes will change.

A new option, "Inherited tags" has also been added, which will affect how nodes are filtered based on the
selected tags. When checked, each node will be considered as having its own tags and the tags of its ancestors.
For example, if we had something like:

 - Node A [#Important]
   - Node B
   - Node C [#Urgent]
 - Node D [#Important][#Urgent]

If checked and we filtered by Urgent AND Important, "Node C" and "Node D" would be selected
If unchecked, then only "Node D" would be selected

This option will be considered not only with the filter associated with "Show use of tags in tree"
(in the Tags tab) but also with the tag filtering performed using the text control in the tree panel.

There is also a new tree context menu option, synchronized with the check in the Tags tab:
"View / Filter | Inherited tags"

* Filtering with the text control in the tree panel has been revised so that the tree is filtered
  as we add new tags
  - Also: Ctrl+Click on this control will clear it

* Fixed: The "View/Filter | Show Children" and "View/Filter | Use Node Path" tree menu options might
  not display the correct value initially.