Change in tag scope identification (also when folding it by pressing Ctrl+Click)
Now, a tag will affect by default the entire paragraph where it is inserted, regardless of whether it is
at the beginning, middle, or end. Therefore, all text in a paragraph such as:

  Blah, blah, blah, ...
  blah, blah .... #ToDo

will be tagged with #ToDo.

If we press Ctrl+Click on #ToDo (with a ":" or immediately before a line break), the folding will also affect this tag,
where this type of "open" writing will modify its scope: it will start at the beginning of the word #ToDo and end at
the first position it finds: ##, #ToDo, or the end of the note.

Some examples:

  #Important
  Bla bla ....

  Other paragraph ....
  ...
  .... ##Important

  This is not important, but from here it is: #Important: Bla Bla ...
  Bla bla ....

  Other paragraph ....
  .... ##

  #Done:
  ...
  ...
  <End of note>